pyproject.toml
["build-system"]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "milvus-service"
version = "1.0.0"
description = "基于 Milvus 的 RAG 模块（operators 与存储封装），可在其他项目中直接复用。"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "BaseVector Core" },
]
keywords = ["milvus", "rag", "vector database", "retrieval", "chunking"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

dependencies = [
  # Milvus & 向量检索相关
  "pymilvus>=2.3.0",
  "numpy>=1.24.0",
  "sentence-transformers>=2.2.0",
  "transformers>=4.35.0",
  "torch>=2.0.0",
  "spacy>=3.7.0",

  # 文档解析相关
  "PyMuPDF>=1.23.0",
  "python-docx>=1.1.0",
  "python-pptx>=0.6.23",
  "beautifulsoup4>=4.12.0",
  "markdown>=3.5.0",
  "Pillow>=10.0.0",
  "pytesseract>=0.3.10",

  # 配置、日志 & 工具
  "pydantic-settings>=2.0.0",
  "PyYAML>=6.0",
  "loguru>=0.7.0",
  "httpx>=0.25.0",
  "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "black",
  "isort",
  "mypy",
]

[project.urls]
Homepage = "https://example.com/"
Repository = "https://example.com/your-repo.git"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["milvus_service*", "ability*"]

